Rename FeedViewModel.kt
