+++ app/src/main/java/com/olderwold/reddit/feature/feed/FeedActivity.kt
@@ -6,16 +6,14 @@ import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
-import com.olderwold.reddit.MainViewModel
 import com.olderwold.reddit.feature.web.WebPage
-import com.olderwold.reddit.ui.element.RedditHotList
 import com.olderwold.reddit.ui.theme.YetanotherredditTheme
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
 @AndroidEntryPoint
 class FeedActivity : ComponentActivity() {
-    private val mainViewModel: MainViewModel by viewModels()
+    private val feedViewModel: FeedViewModel by viewModels()
 
     @Inject
     internal lateinit var webPage: WebPage
@@ -25,7 +23,7 @@ class FeedActivity : ComponentActivity() {
         setContent {
             YetanotherredditTheme {
                 Surface(color = MaterialTheme.colors.background) {
-                    RedditHotList(mainViewModel, webPage)
+                    RedditHotList(feedViewModel, webPage)
                 }
             }
         }
