+++ app/src/main/java/com/olderwold/reddit/feature/feed/FeedViewModel.kt
@@ -1,15 +1,14 @@
-package com.olderwold.reddit
+package com.olderwold.reddit.feature.feed
 
 import androidx.lifecycle.ViewModel
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
-import com.olderwold.reddit.feature.feed.FeedPagingSource
 import com.olderwold.reddit.feature.feed.domain.FeedItem
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-internal class MainViewModel @Inject constructor(
+internal class FeedViewModel @Inject constructor(
     private val pagingSource: FeedPagingSource,
 ) : ViewModel() {
     val pager: Pager<String, FeedItem>
